インターン二日目 + 土曜の進捗。 ・ジャスト回避の演出 ・ダメージ演出 ・弾を撃つ敵の移動速度上昇 ・敵弾の視認性向上、追尾精度を調整 ・ビームの太さを上昇

+++ Project/GameObject/Player/PlayerBullet.h
@@ -3,6 +3,7 @@
 #include "CollisionManager.h"
 #include "Object3D.h"
 #include "WorldTransform.h"
+#include "GameTime.h"
 
 class IEnemy;
 class Player;
@@ -78,7 +79,7 @@ private: // メンバ変数
 	// 寿命<frm>
 	static const int32_t kLifeTime = 60 * 1000;
 	// 死亡タイマー
-	int32_t deathTimer_ = kLifeTime;
+	float deathTimer_ = kLifeTime;
 	// 死亡フラグ
 	bool isDead_ = false;
 };
